requirements.txt
# Core LLM libraries
openai>=1.0.0
anthropic>=0.7.0
huggingface-hub>=0.16.0
transformers>=4.30.0
torch>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

